commentedcursor: beginnings
Start by moving all the cursor related helpers to a separate file
